+++ pkgs/intl4x/lib/src/collator/collator_impl.dart
@@ -2,9 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../../intl.dart';
+import '../../intl4x.dart';
 
-import '../../intl_testutils.dart';
+import '../../intl4x_testutils.dart';
 import 'collator_options.dart';
 
 abstract class Collator {
