+++ pkgs/intl4x/lib/src/datetime_format/datetime_format_4x.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../../intl.dart';
+import '../../intl4x.dart';
 import 'datetime_format_options.dart';
 import 'datetime_formatter.dart';
