+++ pkgs/intl4x/lib/src/number_format/number_format.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../../intl.dart';
+import '../../intl4x.dart';
 
 import 'number_format_4x.dart';
 import 'number_format_options.dart';
