+++ pkgs/intl4x/lib/src/number_format/number_format_ecma.dart
@@ -5,7 +5,7 @@
 import 'package:js/js.dart';
 import 'package:js/js_util.dart';
 
-import '../../intl.dart';
+import '../../intl4x.dart';
 @JS()
 import '../utils.dart';
 import 'number_format_options.dart';
