+++ pkgs/intl4x/lib/src/number_format/number_formatter.dart
@@ -2,9 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../../intl.dart';
+import '../../intl4x.dart';
 
-import '../../intl_testutils.dart';
+import '../../intl4x_testutils.dart';
 import 'number_format_options.dart';
 
 abstract class NumberFormatter {
